dom/media/tests/mochitest/test_peerConnection_setLocalOfferInHaveRemoteOffer.html

<!DOCTYPE HTML>
<html>
<head>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript" src="mediaStreamPlayback.js"></script>
  <script type="application/javascript" src="pc.js"></script>
  <script type="application/javascript" src="templates.js"></script>
  <script type="application/javascript" src="turnConfig.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript">
  createHTML({
    bug: "784519",
    title: "setLocalDescription (offer) in 'have-remote-offer'"
  });

var test;
runNetworkTest(function () {
  test = new PeerConnectionTest();
  test.setMediaConstraints([{audio: true}], [{audio: true}]);
  test.chain.removeAfter("PC_REMOTE_SET_REMOTE_DESCRIPTION");

  test.chain.append([
    function PC_REMOTE_SET_LOCAL_OFFER(test) {
      test.pcRemote.setLocalDescriptionAndFail(test.pcLocal._latest_offer)
        .then(err => {
          is(err.name, "InvalidStateError", "Error is InvalidStateError");
        });
    }
  ]);

  test.run();
});
</script>
</pre>
</body>
</html>
